+++ blindtab-react/src/utils/db.ts
@@ -12,7 +12,8 @@ import {
   arrayRemove,
   query,
   where,
-  Timestamp
+  Timestamp,
+  serverTimestamp
 } from 'firebase/firestore';
 
 // Test function to create a sample song
@@ -50,9 +51,10 @@ export const songOperations = {
     try {
       // Test connection by trying to get a document
       await getDocs(collection(db, COLLECTIONS.SONGS));
+      console.log('[Firestore] Successfully initialized');
       return true;
     } catch (error) {
-      console.error('Failed to initialize Firestore:', error);
+      console.error('[Firestore] Failed to initialize:', error);
       return false;
     }
   },
@@ -60,105 +62,128 @@ export const songOperations = {
   checkConnection: async (): Promise<boolean> => {
     try {
       await getDocs(collection(db, COLLECTIONS.SONGS));
+      console.log('[Firestore] Connection check successful');
       return true;
     } catch (error) {
-      console.error('Failed to connect to Firestore:', error);
+      console.error('[Firestore] Connection check failed:', error);
       return false;
     }
   },
 
   getAllSongs: async (): Promise<Song[]> => {
     try {
+      console.log('[Firestore] Fetching all songs');
       const snapshot = await getDocs(collection(db, COLLECTIONS.SONGS));
       return snapshot.docs.map(doc => ({
         id: doc.id,
         ...doc.data()
       } as Song));
     } catch (error) {
-      console.error('Failed to get songs:', error);
+      console.error('[Firestore] Failed to get songs:', error);
       return [];
     }
   },
 
   getSongById: async (id: string): Promise<SongData | null> => {
     try {
+      console.log(`[Firestore] Fetching song ${id}`);
       const docRef = doc(db, COLLECTIONS.SONGS, id);
       const docSnap = await getDoc(docRef);
       
       if (docSnap.exists()) {
         return docSnap.data() as SongData;
       }
+      console.log(`[Firestore] Song ${id} not found`);
       return null;
     } catch (error) {
-      console.error(`Failed to get song ${id}:`, error);
+      console.error(`[Firestore] Failed to get song ${id}:`, error);
       return null;
     }
   },
 
   createSong: async (songData: SongData): Promise<string | null> => {
     try {
+      console.log('[Firestore] Creating new song');
       const docRef = doc(collection(db, COLLECTIONS.SONGS));
-      await setDoc(docRef, songData);
+      const timestamp = serverTimestamp();
+      
+      await setDoc(docRef, {
+        ...songData,
+        createdAt: timestamp,
+        updatedAt: timestamp
+      });
+
+      console.log(`[Firestore] Created song with ID: ${docRef.id}`);
       return docRef.id;
     } catch (error) {
-      console.error('Failed to create song:', error);
+      console.error('[Firestore] Failed to create song:', error);
       return null;
     }
   },
 
   updateSong: async (id: string, markdown: string): Promise<boolean> => {
     try {
+      console.log(`[Firestore] Updating song ${id}`);
       const docRef = doc(db, COLLECTIONS.SONGS, id);
-      await updateDoc(docRef, { markdown });
+      await updateDoc(docRef, {
+        markdown,
+        updatedAt: serverTimestamp()
+      });
       return true;
     } catch (error) {
-      console.error(`Failed to update song ${id}:`, error);
+      console.error(`[Firestore] Failed to update song ${id}:`, error);
       return false;
     }
   },
 
   deleteSong: async (id: string): Promise<boolean> => {
     try {
+      console.log(`[Firestore] Deleting song ${id}`);
       await deleteDoc(doc(db, COLLECTIONS.SONGS, id));
       return true;
     } catch (error) {
-      console.error(`Failed to delete song ${id}:`, error);
+      console.error(`[Firestore] Failed to delete song ${id}:`, error);
       return false;
     }
   },
 
   addTagToSong: async (songId: string, tagName: string): Promise<boolean> => {
     try {
+      console.log(`[Firestore] Adding tag ${tagName} to song ${songId}`);
       const docRef = doc(db, COLLECTIONS.SONGS, songId);
       await updateDoc(docRef, {
-        tags: arrayUnion(tagName)
+        tags: arrayUnion(tagName),
+        updatedAt: serverTimestamp()
       });
       return true;
     } catch (error) {
-      console.error(`Failed to add tag ${tagName} to song ${songId}:`, error);
+      console.error(`[Firestore] Failed to add tag ${tagName} to song ${songId}:`, error);
       return false;
     }
   },
 
   removeTagFromSong: async (songId: string, tagName: string): Promise<boolean> => {
     try {
+      console.log(`[Firestore] Removing tag ${tagName} from song ${songId}`);
       const docRef = doc(db, COLLECTIONS.SONGS, songId);
       await updateDoc(docRef, {
-        tags: arrayRemove(tagName)
+        tags: arrayRemove(tagName),
+        updatedAt: serverTimestamp()
       });
       return true;
     } catch (error) {
-      console.error(`Failed to remove tag ${tagName} from song ${songId}:`, error);
+      console.error(`[Firestore] Failed to remove tag ${tagName} from song ${songId}:`, error);
       return false;
     }
   },
 
   getAllTags: async () => {
     try {
+      console.log('[Firestore] Fetching all tags');
       const snapshot = await getDocs(collection(db, COLLECTIONS.TAGS));
       return snapshot.docs.map(doc => doc.data().name);
     } catch (error) {
-      console.error('Failed to get tags:', error);
+      console.error('[Firestore] Failed to get tags:', error);
       return [];
     }
   }
