+++ firestore.rules
@@ -1,0 +1,26 @@
+rules_version = '2';
+
+service cloud.firestore {
+  match /databases/{database}/documents {
+    // Allow read/write access during development
+    match /{document=**} {
+      allow read, write: if true;
+    }
+
+    // TODO: Add proper security rules for production
+    match /songs/{songId} {
+      allow read: if true;  // Anyone can read songs
+      allow write: if false; // No writes for now
+    }
+
+    match /config/{configId} {
+      allow read: if true;
+      allow write: if false;
+    }
+
+    match /tags/{tagId} {
+      allow read: if true;
+      allow write: if false;
+    }
+  }
+}
